MPI: Add stack-local limb storage to MPI objects.
This allows for a small chunk of memory to be added to each mbedtls_mpi
structure instantiation, the amount of which is adjustable via a new
compile-time parameter MBEDTLS_MPI_LOCAL_LIMB_SIZE. If a given MPI can
fit wholly within this chunk of memory (which will typically reside on
the stack), no heap allocation calls for that MPI will be made. Should,
however, an MPI require more space or grow beyond the extents provided
by that chunk, then an allocation call will be made as usual and limb
storage will be diverted appropriately. The default size of stack-local
storage to be reserved is 32 bytes (i.e., 256 bits).

This is particularly useful for elliptic curve operations: a 10%
reduction to execution time was observed for operations against the
secp256{r,k}1 curves with a minor (temporary) increase in the
required amount of stack space. To maintain the current behavior of
always using heap allocation regardless of MPI size, the library can
be compiled with MBEDTLS_MPI_LOCAL_LIMB_SIZE defined to 0.